+++ feature/home/src/main/java/com/arifwidayana/home/domain/home/CategoryProductUseCase.kt
@@ -1,9 +1,9 @@
-package com.arifwidayana.home.domain
+package com.arifwidayana.home.domain.home
 
 import com.arifwidayana.core.base.BaseUseCase
 import com.arifwidayana.core.wrapper.ViewResource
 import com.arifwidayana.home.data.network.repository.HomeRepository
-import com.arifwidayana.shared.data.network.model.mapper.home.CategoryListProductMapper
+import com.arifwidayana.shared.data.network.model.mapper.home.CategoryProductMapper
 import com.arifwidayana.shared.data.network.model.response.home.category.CategoryParamResponse
 import com.arifwidayana.shared.utils.ext.suspendSource
 import kotlinx.coroutines.CoroutineDispatcher
@@ -18,13 +18,13 @@ class CategoryProductUseCase(
 ) : BaseUseCase<Nothing, CategoryDataResource>(coroutineDispatcher) {
     override suspend fun execute(param: Nothing?): Flow<ViewResource<CategoryDataResource>> = flow {
         emit(ViewResource.Loading())
-        homeRepository.categoryProduct().collect { source ->
-            source.suspendSource(
-                doOnSuccess = {
-                    emit(ViewResource.Success(CategoryListProductMapper.toViewParam(it.payload)))
+        homeRepository.categoryProduct().collect {
+            it.suspendSource(
+                doOnSuccess = { source ->
+                    emit(ViewResource.Success(CategoryProductMapper.toViewParam(source.payload)))
                 },
-                doOnError = {
-                    emit(ViewResource.Error(it.exception))
+                doOnError = { error ->
+                    emit(ViewResource.Error(error.exception))
                 }
             )
         }
